Moved SparkInstance to Spark.Api
