--- src/main/java/spark/SparkBase.java
@@ -1,189 +1,0 @@
-package spark;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import spark.exception.ExceptionMapper;
-import spark.route.SimpleRouteMatcher;
-
-/**
- * Spark base class
- */
-public abstract class SparkBase {
-
-    private static final Logger LOG = LoggerFactory.getLogger("spark.Spark");
-    protected static SparkInstance instance = new SparkInstance();
-
-    /**
-     * Set the IP address that Spark should listen on. If not called the default
-     * address is '0.0.0.0'. This has to be called before any route mapping is
-     * done.
-     *
-     * @param ipAddress The ipAddress
-     * @deprecated replaced by {@link #ipAddress(String)}
-     */
-    public static synchronized void setIpAddress(String ipAddress) {
-        ipAddress(ipAddress);
-    }
-
-    /**
-     * Set the IP address that Spark should listen on. If not called the default
-     * address is '0.0.0.0'. This has to be called before any route mapping is
-     * done.
-     *
-     * @param ipAddress The ipAddress
-     */
-    public static synchronized void ipAddress(String ipAddress) {
-        instance.ip(ipAddress);
-    }
-
-    /**
-     * Set the port that Spark should listen on. If not called the default port
-     * is 4567. This has to be called before any route mapping is done.
-     * If provided port = 0 then the an arbitrary available port will be used.
-     *
-     * @param port The port number
-     * @deprecated replaced by {@link #port(int)}
-     */
-    public static synchronized void setPort(int port) {
-        port(port);
-    }
-
-    /**
-     * Set the port that Spark should listen on. If not called the default port
-     * is 4567. This has to be called before any route mapping is done.
-     * If provided port = 0 then the an arbitrary available port will be used.
-     *
-     * @param port The port number
-     */
-    public static synchronized void port(int port) {
-        instance.port(port);
-    }
-
-    /**
-     * Set the connection to be secure, using the specified keystore and
-     * truststore. This has to be called before any route mapping is done. You
-     * have to supply a keystore file, truststore file is optional (keystore
-     * will be reused).
-     * This method is only relevant when using embedded Jetty servers. It should
-     * not be used if you are using Servlets, where you will need to secure the
-     * connection in the servlet container
-     *
-     * @param keystoreFile       The keystore file location as string
-     * @param keystorePassword   the password for the keystore
-     * @param truststoreFile     the truststore file location as string, leave null to reuse
-     *                           keystore
-     * @param truststorePassword the trust store password
-     * @deprecated replaced by {@link #secure(String, String, String, String)}
-     */
-    public static synchronized void setSecure(String keystoreFile,
-                                              String keystorePassword,
-                                              String truststoreFile,
-                                              String truststorePassword) {
-        secure(keystoreFile, keystorePassword, truststoreFile, truststorePassword);
-    }
-
-    /**
-     * Set the connection to be secure, using the specified keystore and
-     * truststore. This has to be called before any route mapping is done. You
-     * have to supply a keystore file, truststore file is optional (keystore
-     * will be reused).
-     * This method is only relevant when using embedded Jetty servers. It should
-     * not be used if you are using Servlets, where you will need to secure the
-     * connection in the servlet container
-     *
-     * @param keystoreFile       The keystore file location as string
-     * @param keystorePassword   the password for the keystore
-     * @param truststoreFile     the truststore file location as string, leave null to reuse
-     *                           keystore
-     * @param truststorePassword the trust store password
-     */
-    public static synchronized void secure(String keystoreFile,
-                                           String keystorePassword,
-                                           String truststoreFile,
-                                           String truststorePassword) {
-        instance.secure(keystoreFile, keystorePassword, truststoreFile, truststorePassword);
-    }
-
-    /**
-     * Sets the folder in classpath serving static files. Observe: this method
-     * must be called before all other methods.
-     *
-     * @param folder the folder in classpath.
-     */
-    public static synchronized void staticFileLocation(String folder) {
-        instance.staticFileLocation(folder);
-    }
-
-    /**
-     * Sets the external folder serving static files. <b>Observe: this method
-     * must be called before all other methods.</b>
-     *
-     * @param externalFolder the external folder serving static files.
-     */
-    public static synchronized void externalStaticFileLocation(String externalFolder) {
-        instance.externalStaticFileLocation(externalFolder);
-    }
-
-    public static synchronized void stop() {
-        instance.stop();
-    }
-
-    static synchronized void runFromServlet() {
-        instance.runFromServlet();
-    }
-
-    /**
-     * Wraps the route in RouteImpl
-     *
-     * @param path  the path
-     * @param route the route
-     * @return the wrapped route
-     */
-    protected static RouteImpl wrap(final String path, final Route route) {
-        return instance.wrap(path, route);
-    }
-
-    /**
-     * Wraps the route in RouteImpl
-     *
-     * @param path       the path
-     * @param acceptType the accept type
-     * @param route      the route
-     * @return the wrapped route
-     */
-    protected static RouteImpl wrap(final String path, String acceptType, final Route route) {
-        return instance.wrap(path, acceptType, route);
-    }
-
-    /**
-     * Wraps the filter in FilterImpl
-     *
-     * @param path   the path
-     * @param filter the filter
-     * @return the wrapped route
-     */
-    protected static FilterImpl wrap(final String path, final Filter filter) {
-        return instance.wrap(path, filter);
-    }
-
-    /**
-     * Wraps the filter in FilterImpl
-     *
-     * @param path       the path
-     * @param acceptType the accept type
-     * @param filter     the filter
-     * @return the wrapped route
-     */
-    protected static FilterImpl wrap(final String path, String acceptType, final Filter filter) {
-        return instance.wrap(path, acceptType, filter);
-    }
-
-    public static ExceptionMapper getExceptionMapper() {
-        return instance.exceptionMapper;
-    }
-
-    public static SimpleRouteMatcher getRouteMatcher() {
-        return instance.routeMatcher;
-    }
-}
